experiments/exp5/requirements_confusion_matrix.txt
# Requirements for Confusion Matrix Generation
# Install with: pip install -r requirements_confusion_matrix.txt

# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine learning
scikit-learn>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.60.0

# Optional: for existing evaluation scripts
openai>=1.0.0
anthropic>=0.18.0
transformers>=4.30.0
torch>=2.0.0
rouge-score>=0.1.2
nltk>=3.6
